// frontend/src/api/generated/api.ts
export interface BudgetUiDTO {
    'id'?: string;
    'name'?: string;
    'description'?: string;
    'isActive'?: boolean;
    'period'?: BudgetUiDTOPeriodEnum;
    'amount'?: number;
    'spentAmount'?: number;
    'periodStart'?: string;
    'periodEnd'?: string;
    'category'?: CategorySimplifiedUiDTO;
}

export const BudgetUiDTOPeriodEnum = {
    Week: 'WEEK',
    Month: 'MONTH',
    Year: 'YEAR'
} as const;

export type BudgetUiDTOPeriodEnum = typeof BudgetUiDTOPeriodEnum[keyof typeof BudgetUiDTOPeriodEnum];
// ...
export interface CategorySimplifiedUiDTO {
    'id'?: string;
    'name'?: string;
}
export interface CategoryUiDTO {
    'id'?: string;
    'name'?: string;
    'description'?: string;
    'isActive'?: boolean;
    'createdAt'?: string;
    'weeklyBudget'?: BudgetUiDTO;
    'monthlyBudget'?: BudgetUiDTO;
    'yearlyBudget'?: BudgetUiDTO;
}
// ...
export interface GetDefaultBudgetsResponseUiDTO {
    'weeklyAmount'?: number;
    'monthlyAmount'?: number;
    'yearlyAmount'?: number;
}
// ...
export interface SetDefaultBudgetsRequestUiDTO {
    'categoryId'?: string;
    'weeklyAmount'?: number;
    'monthlyAmount'?: number;
    'yearlyAmount'?: number;
}